Rename to scalgoprotoc and use as Python package
