+++ scalgoprotoc/error.py
