+++ scalgoprotoc/magic.py
