+++ scalgoprotoc/util.py
